rename EdSchema
