+++ src/inputrules/ed-input-rules.js
@@ -4,12 +4,12 @@ import {inputRules, allInputRules} from 'prosemirror-inputrules'
 import {buildInputRules, buildKeymap} from 'prosemirror-example-setup'
 import {keymap} from 'prosemirror-keymap'
 import {baseKeymap} from 'prosemirror-commands'
-import EdSchemaFull from '../schema/ed-schema-full'
+import EdSchema from '../schema/ed-schema'
 // import inputCode from './input-code'
 // import iosDoubleSpace from './ios-double-space'
 
 
-const edRules = buildInputRules(EdSchemaFull)
+const edRules = buildInputRules(EdSchema)
 
 // edRules.push(inputCode)
 
@@ -23,4 +23,4 @@ const rules = allInputRules.concat(edRules)
 export const edInputRules = inputRules({rules})
 
 export const edBaseKeymap = keymap(baseKeymap)
-export const edKeymap = keymap(buildKeymap(EdSchemaFull))
+export const edKeymap = keymap(buildKeymap(EdSchema))
