+++ src/inputrules/ed-keymap.js
@@ -1,5 +1,5 @@
 import {buildKeymap} from 'prosemirror-example-setup'
-import EdSchema from '../schema/ed-schema-full'
+import EdSchema from '../schema/ed-schema'
 
 
 const EdKeymap = buildKeymap(EdSchema)
